Added a practice page for rearranging code solutions.
Currently only the most recent submission for the the logged in user is
presented. Simple rearranging is implemented by importing the
sortable.js library through a cdn. Other options were to serve locally
through the static folder or open up the nodejs package management
can of worms, which seemed time consuming and not worth it as long as
the project requirements are small.