Added answer checking to the practice page
The user's permutation of the code solution is submitted to
the server through a post request (using the fetch api). The
server checks whether the permutation matches the saved solution
and returns whether the submission was correct or not.

All this is current written inline in the html template, should be
refactored at some point.