Fixed bug with checking answers on practice pages that have an id
When checking answers, the javascript attached to the submit button
sends the request to the request path + "/check", so the paths are
actually different when viewing a practice that may or may not have
an id in the url. Solution is to attach both routes the check_answer()
method in the practice blueprint.